requirements.txt
streamlit==1.28.0
torch==2.1.0
transformers==4.35.0
sentence-transformers==2.2.2
langchain==0.1.0
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0
plotly==5.17.0
